Move runRules to utils.js
