+++ src/utils.js
@@ -49,5 +49,53 @@ module.exports = {
       }
     }
     return null;
+  },
+
+  runRules: function (value, currentValues, validations, validationRules) {
+    const results = {
+      errors: [],
+      failed: [],
+      success: [],
+    };
+
+    if (Object.keys(validations).length) {
+      Object.keys(validations).forEach((validationMethod) => {
+        if (validationRules[validationMethod] && typeof validations[validationMethod] === 'function') {
+          throw new Error(`Formsy does not allow you to override default validations: ${validationMethod}`);
+        }
+
+        if (!validationRules[validationMethod] && typeof validations[validationMethod] !== 'function') {
+          throw new Error(`Formsy does not have the validation rule: ${validationMethod}`);
+        }
+
+        if (typeof validations[validationMethod] === 'function') {
+          const validation = validations[validationMethod](currentValues, value);
+          if (typeof validation === 'string') {
+            results.errors.push(validation);
+            results.failed.push(validationMethod);
+          } else if (!validation) {
+            results.failed.push(validationMethod);
+          }
+          return;
+        } else if (typeof validations[validationMethod] !== 'function') {
+          const validation = validationRules[validationMethod](
+            currentValues, value, validations[validationMethod],
+          );
+          if (typeof validation === 'string') {
+            results.errors.push(validation);
+            results.failed.push(validationMethod);
+          } else if (!validation) {
+            results.failed.push(validationMethod);
+          } else {
+            results.success.push(validationMethod);
+          }
+          return;
+        }
+
+        results.success.push(validationMethod);
+      });
+    }
+
+    return results;
   }
 };
